Dev (#93)
* added: querying warehouses with locations and by site

* Enhance Verifactu integration and localization support (#92)

* Create table salesinvoicehash

* Fix DB mistakes

* refactor: sales invoice verifactu request table

* checkpoint

* deleted: redundant service

* added: invoice hash generator

* added: background service to send the invoices to verifactu

* fix merge

* changed: invoice serie to shared entity

* added: notification to verifactu with hash calculation

* fixed: verifactu hash generation

* added: find verifactu requests by year / month

* refactor: improve best practices on Verifactu library

* refactor: divide services into SalesInvoiceService & VerifactuService

* added: QR code generation on Verifactu invoice integration

* added: qr code URL separate setting

* change qr url field

* added: QrCodeReportTag to InvoiceReportDto

* Add localization support and middleware integration

Updated project files to include Microsoft.Extensions.Localization for localization. Added JSON resource files and created a LocalizationController for testing. Implemented CultureMiddleware for culture detection and introduced a custom JsonStringLocalizer. Refactored LifecycleController to use localized messages and established a LocalizationService for string retrieval. Configured localization services in Program.cs and defined an ILocalizationService interface for standardization.

* refactor services and controllers to add translation keys with localitzation service

* changed: json format with culture property

* move in required directory

* added: better controller tests

* Add Language management functionality

This commit introduces a new `Language` entity along with its repository, service, and controller to manage language data within the application. The `LanguageController` handles API requests for retrieving, creating, updating, and deleting languages. The `ILanguageService` interface and `LanguageRepository` provide methods for language operations. A migration is added to create the `Languages` table in the database with default entries. Additionally, the model snapshot and entity configuration are updated to include the new `Language` entity, enhancing the application's language management capabilities.

* more...

* added: multilanguage per user support

* removed: translation keys

* fixed: refresh token capability

* added: reference dependency messages to localitzaton files

* refactor

* added: static translation keys for reporting

* use primary constructor

* Add Verifactu integration and refactor invoice handling

This commit introduces a new `Verifactu` project with controllers for handling invoice requests, including cancellation and querying. The `AppSettings` class was added for configuration management, and existing controllers were refactored to utilize a unit of work pattern.

New properties were added to the `SalesInvoiceVerifactuRequest` entity for QR code handling, and migrations were created to update the database schema accordingly. The `HashGenerator` and `VerifactuFormatUtils` classes were introduced for generating hashes and formatting data.

Additionally, the project file was updated to include necessary references and settings for the new functionality.

* fixed: degugging on .net8 compilation

* added: final report translation

* refactor: create specific entity reporting service and add multilanguage texts by customer preferred language

* refactor: compilation warnings

* fixed: remove unnecessary characters

* added: translation keys for verifactu integration
added: endpoints to get pending invoices and cancel an integrated invoice

* feat: enhance Verifactu integration with new endpoints for date range queries and improve invoice retrieval logic

* fixed: localitzation service as singleton

* fixed: request integration cutoff by invoice date

* changed: special caracters on strings because bad encoding

* added: default site on enterprise

* added: due dates recreation

* fixed: staging detected bugs

* fix: update invoice import retrieval to use unit of work for better data access

* fix: localize delivery note header in invoice report

---------

Co-authored-by: Marc Gurt <marc.gurt@mapex.io>
Co-authored-by: marcgurt-mapex <167959703+marcgurt-mapex@users.noreply.github.com>